// src/generated/models/Cfm_tmshandoffsModel.ts
/*!
 * Copyright (C) Microsoft Corporation. All rights reserved.
 * This file is autogenerated. Do not edit this file directly.
 */
export const Cfm_tmshandoffscfm_decisionfromwhere = {
  1: 'Cash Flow',
  2: 'AP to Treasury  Req',
  4: 'AP to AP',
  3: 'AP to Treasury  Submit ',
  5: 'AR'
} as const;
export type Cfm_tmshandoffscfm_decisionfromwhere = keyof typeof Cfm_tmshandoffscfm_decisionfromwhere;
export const Cfm_tmshandoffscfm_priority = {
  123200000: 'Low',
  123200001: 'Medium',
  123200002: 'High',
  931940001: 'Critical'
} as const;
export type Cfm_tmshandoffscfm_priority = keyof typeof Cfm_tmshandoffscfm_priority;
export const Cfm_tmshandoffscfm_progress = {
  100000001: 'In Progress ',
  100000005: 'On Hold',
  123200002: 'Closed',
  123200003: 'Cancelled',
  123200004: 'New',
  123200005: 'Submited',
  931940001: 'Rejected'
} as const;
export type Cfm_tmshandoffscfm_progress = keyof typeof Cfm_tmshandoffscfm_progress;
export const Cfm_tmshandoffscfm_tagename = {
  766340000: 'Collection',
  766340001: 'AR',
  766340002: 'AP'
} as const;
export type Cfm_tmshandoffscfm_tagename = keyof typeof Cfm_tmshandoffscfm_tagename;
export const Cfm_tmshandoffsstatecode = {
  0: 'Active',
  1: 'Inactive'
} as const;
export type Cfm_tmshandoffsstatecode = keyof typeof Cfm_tmshandoffsstatecode;
export const Cfm_tmshandoffsstatuscode = {
  1: 'Active',
  2: 'Inactive'
} as const;
export type Cfm_tmshandoffsstatuscode = keyof typeof Cfm_tmshandoffsstatuscode;

export interface Cfm_tmshandoffsBase {
  "cfm_Assignee@odata.bind"?: string;
  cfm_decision?: string;
  cfm_decisionfromwhere?: Cfm_tmshandoffscfm_decisionfromwhere;
  cfm_duedate?: string;
  cfm_priority?: Cfm_tmshandoffscfm_priority;
  cfm_progress?: Cfm_tmshandoffscfm_progress;
  cfm_sla?: string;
  cfm_tagename?: Cfm_tmshandoffscfm_tagename;
  cfm_taskdescription?: string;
  "cfm_TasksTMS@odata.bind"?: string;
  cfm_tasktitle?: string;
  cfm_tmshandoffid: string;
  cfm_typeaction?: string;
  importsequencenumber?: number;
  overriddencreatedon?: string;
  statecode: Cfm_tmshandoffsstatecode;
  statuscode?: Cfm_tmshandoffsstatuscode;
  timezoneruleversionnumber?: number;
  utcconversiontimezonecode?: number;
}

export interface Cfm_tmshandoffs extends Cfm_tmshandoffsBase {
  cfm_assigneename?: string;
  cfm_assigneeyominame?: string;
  cfm_decisionfromwherename?: string;
  cfm_priorityname?: string;
  cfm_progressname?: string;
  cfm_tagenamename?: string;
  cfm_taskstmsname?: string;
  cfm_tmscode?: string;
  createdbyname?: string;
  createdbyyominame: string;
  createdon?: string;
  createdonbehalfbyname?: string;
  createdonbehalfbyyominame: string;
  modifiedbyname?: string;
  modifiedbyyominame: string;
  modifiedon?: string;
  modifiedonbehalfbyname?: string;
  modifiedonbehalfbyyominame: string;
  ownerid: string;
  owneridname: string;
  owneridtype: string;
  owneridyominame: string;
  owningbusinessunitname: string;
  statecodename?: string;
  statuscodename?: string;
  versionnumber?: number;
  cfm_assignee?: object;
  _cfm_assignee_value?: string;
  cfm_taskstms?: object;
  _cfm_taskstms_value?: string;
  createdby?: object;
  _createdby_value?: string;
  createdonbehalfby?: object;
  _createdonbehalfby_value?: string;
  modifiedby?: object;
  _modifiedby_value?: string;
  modifiedonbehalfby?: object;
  _modifiedonbehalfby_value?: string;
  owningbusinessunit?: object;
  _owningbusinessunit_value?: string;
  owningteam?: object;
  _owningteam_value?: string;
  owninguser?: object;
  _owninguser_value?: string;
}